backend/requirements.txt
# Django Core (5.2 LTS - Long Term Support)
Django>=5.2,<5.3
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Authentication
django-allauth>=0.61,<1.0
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Redis & Celery
redis>=5.0,<6.0
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Security & Encryption
cryptography>=42.0,<43.0
python-dotenv>=1.0,<2.0

# HTTP
requests>=2.31,<3.0
requests-oauthlib>=1.3,<2.0

# OAuth Providers
google-auth>=2.28,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.118,<3.0
msal>=1.26,<2.0

# Utilities
Pillow>=10.2,<11.0
python-dateutil>=2.8,<3.0
pytz>=2024.1
openpyxl>=3.1,<4.0

# Development
ipython>=8.20,<9.0
